Move math utilities to one folder
