+++ src/pygerber/backend/abstract/draw_commands/draw_bounding_box.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 from functools import cached_property
 
 from pygerber.backend.abstract.backend_cls import Backend
-from pygerber.backend.abstract.bounding_box import BoundingBox
 from pygerber.backend.abstract.draw_commands.draw_command import DrawCommand
-from pygerber.backend.abstract.offset import Offset
+from pygerber.gerberx3.math.bounding_box import BoundingBox
+from pygerber.gerberx3.math.offset import Offset
 from pygerber.gerberx3.state_enums import Polarity
