+++ src/pygerber/backend/abstract/draw_commands/draw_region.py
@@ -4,10 +4,10 @@ from __future__ import annotations
 from functools import cached_property
 
 from pygerber.backend.abstract.backend_cls import Backend
-from pygerber.backend.abstract.bounding_box import BoundingBox
 from pygerber.backend.abstract.draw_commands.draw_command import DrawCommand
-from pygerber.backend.abstract.offset import Offset
-from pygerber.backend.abstract.vector_2d import Vector2D
+from pygerber.gerberx3.math.bounding_box import BoundingBox
+from pygerber.gerberx3.math.offset import Offset
+from pygerber.gerberx3.math.vector_2d import Vector2D
 from pygerber.gerberx3.state_enums import Polarity
