+++ src/pygerber/backend/abstract/drawing_target.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Optional, TypeVar
 
-from pygerber.backend.abstract.bounding_box import BoundingBox
-from pygerber.backend.abstract.vector_2d import Vector2D
+from pygerber.gerberx3.math.bounding_box import BoundingBox
+from pygerber.gerberx3.math.vector_2d import Vector2D
 
 if TYPE_CHECKING:
     from types import TracebackType
