+++ src/pygerber/gerberx3/math/__init__.py
@@ -1,0 +1,1 @@
+"""Module containing math utilities."""
