+++ src/pygerber/gerberx3/math/offset.py
